+++ R/read.am.r
@@ -1,13 +1,3 @@
-# Input/Output functions of the AMOR package are here
-read.pool <- function(file){
-    # Function that takes a text file pooling
-    # map and reads it into R.
-    fc <- file(file)
-    Pool <- strsplit(readLines(fc), "\t")
-    close(fc)
-    return(Pool)
-}
-
 #' Read abundance matrix
 #' 
 #' Reads an abundance matrix into a Dataset object
@@ -38,7 +28,7 @@ read.pool <- function(file){
 #' If simplify = TRUE, it returns a numeric abundance matrix. Any taxonomic
 #' information is lost in this case.
 #' 
-#' @author Sur from Dangl Lab.
+#' @author Sur Herrera Paredes
 #' 
 #' @seealso \code{\link{create_dataset}}
 #' 
@@ -71,23 +61,4 @@ read.am <- function(file,format = "am", taxonomy = FALSE, simplify = FALSE){
   }
   
   return(Dat)
-}
-
-write.qiime <- function(...) UseMethod("write.qiime")
-
-write.qiime.default <- function(Tab,file){
-    # Function that takes an OTU table, and writes a file in
-    # QIIME format with it.
-    first <- "# QIIME v1.3.0 OTU table"
-    header <- colnames(Tab)
-    header <- c("#OTU ID", header)
-    header <- paste(header,collapse="\t")
-    fileConn <- file(file)
-    writeLines(c(first,header),fileConn,sep="\n")
-    close(fileConn)
-    write.table(Tab,file=file,col.names=F,row.names=T,sep="\t",quote=F,append=T)
-}
-
-write.qiime.Dataset <- function(Dat,file){
-  write.qiime.default(Tab = Dat$Tab, file = file)
 }