+++ NEWS.md
@@ -7,6 +7,8 @@ since rarefaction is not recomended to deal with uneven sampling.
 5. Eliminated summaizeOTU distribution. Never implemented
 6. Removed normalizeSample function.
 7. Roxyfen documentation and examples for get_tax_level
+8. Removed read.pool function
+9. Documentation and examples for write.qiime
 
 # 0.1-3
 1. Added support in subset.Dataset for when only one sample remains